working

+++ bin/menuV4.sh
@@ -17,7 +17,7 @@ display_help() {
     echo "   ctrl-p                 select playbook"
     echo "   ctrl-h                 select host(s)"
     echo "   ctrl-t                 select tags"
-    echo "   ctrl-r                 select task (role)"
+    echo "   ctrl-r                 select task"
     echo "   ctrl-x                 execute playbook"
     echo "   ctrl-c                 exit"
     echo
@@ -49,6 +49,8 @@ selected="#c678dd"
 preview_bg="#21252b"
 preview_fg="#98c379"
 
+declare -rx GUM_SPIN_SPINNER="pulse"
+
 # Variables to store selections
 selected_playbook=""
 selected_hosts=""
@@ -72,7 +74,9 @@ select_playbook() {
     selected_playbook=$(find playbooks/ -name "*.yml" | \
         fzf --preview 'bat --style=numbers --color=always {}' \
             --preview-window=right:60% \
+            --layout="reverse" \
             --header="Select Playbook" \
+            --tac \
             --prompt="Playbook > ")
 }
 
@@ -80,6 +84,7 @@ select_playbook() {
 select_hosts() {
     selected_hosts=$(awk '/^\[/ {gsub(/[\[\]]/, ""); print}' inventory.ini | \
         fzf --multi \
+            --layout="reverse" \
             --header="Select Host(s)" \
             --prompt="Hosts > ")
 }
@@ -87,9 +92,10 @@ select_hosts() {
 # Function to select tags
 select_tags() {
     if [[ -n "$selected_playbook" ]]; then
-        selected_tags=$(ansible-playbook "$selected_playbook" --list-tags | \
+        selected_tags=$(ansible-playbook -i inventory.ini "$selected_playbook" --list-tags | \
             grep "TASK TAGS" | cut -d':' -f2 | tr '[' ' ' | tr ']' ' ' | tr ',' '\n' | \
             fzf --multi \
+                --layout="reverse" \
                 --header="Select Tags" \
                 --prompt="Tags > ")
     else
@@ -97,24 +103,37 @@ select_tags() {
     fi
 }
 
-# Function to select task (role)
+# Function to select task
 select_task() {
-    # Create a temporary file to store all tasks
-    temp_file=$(mktemp)
-    find roles/ -type f -name "*.yml" -exec bash -c 'extract_tasks "$0"' {} \; > "$temp_file"
+    if [[ -n "$selected_playbook" ]]; then
+        # Create a temporary file to store all tasks
+        temp_file=$(mktemp)
+        find roles/ -type f -wholename "**/tasks/*.yml" -exec bash -c 'extract_tasks "$0"' {} \; > "$temp_file"
 
-    selected_task=$(cat "$temp_file" | \
-        fzf --preview 'line=$(echo {} | cut -d: -f2); file=$(echo {} | cut -d: -f1); bash -c "preview_task \"$file\" \"$line\""' \
-            --preview-window=right:60% \
-            --header="Select Task" \
-            --prompt="Task > ")
+        selected_task=$(cat "$temp_file" | \
+            fzf --preview 'line=$(echo {} | cut -d: -f2); file=$(echo {} | cut -d: -f1); bash -c "preview_task \"$file\" \"$line\""' \
+                --preview-window=right:50% \
+                --tac \
+                --layout="reverse" \
+                --header="Select Task" \
+                --prompt="Task > ")
 
-    # Remove the temporary file
-    rm "$temp_file"
+        # Remove the temporary file
+        rm "$temp_file"
 
-    if [[ -n "$selected_task" ]]; then
-        selected_task=$(echo "$selected_task" | cut -d: -f3-)
+        if [[ -n "$selected_task" ]]; then
+            selected_task=$(echo "$selected_task" | cut -d: -f3-)
+        fi
+    else
+        echo "Please select a playbook first."
     fi
+}
+
+reset_vars() {
+  unset selected_playbook
+  unset selected_hosts
+  unset selected_tags
+  unset selected_task
 }
 
 # Function to execute playbook
@@ -124,7 +143,7 @@ execute_playbook() {
         return
     fi
 
-    command="ansible-playbook -C $selected_playbook"
+    command="ansible-playbook -C -i inventory.ini $selected_playbook"
 
     if [[ -n "$selected_hosts" ]]; then
         command+=" --limit $selected_hosts"
@@ -132,24 +151,54 @@ execute_playbook() {
 
     if [[ -n "$selected_tags" ]]; then
         command+=" --tags $selected_tags"
-    else
-        # If no specific tags are selected, always include 'always' tag
-        command+=" --tags always"
+    # else
+    #     command+=" --tags always"
     fi
 
     if [[ -n "$selected_task" ]]; then
-        # Instead of --start-at-task, we'll use tags
-        task_tag=$(echo "$selected_task" | tr ' ' '_' | tr '[:upper:]' '[:lower:]')
-        command+=" --tags always,setup,$task_tag"
+        command+=" --start-at-task '$selected_task'"
     fi
 
-    # Always gather facts
     command+=" -e 'gather_facts=true'"
 
-    echo "Executing: $command"
-    eval "$command"
-}
+    # Display confirmation with gum
+    echo "Selected options:"
+    echo "Playbook: $selected_playbook"
+    echo "Hosts: $selected_hosts"
+    echo "Tags: $selected_tags"
+    echo "Task: $selected_task"
+    echo
+    echo "Command to execute:"
+    echo "$command"
+    echo
 
+    if gum confirm "Do you want to execute this playbook?"; then
+        echo "Executing: $command"
+        echo "Please wait while the playbook runs..."
+
+        # # Execute the command and pipe output to gum pager
+        # eval "$command" 2>&1 | gum pager
+        # Execute the command and pipe output to Ruby script
+        eval "$command" 2>&1 | ruby -e '
+            STDIN.each_line do |line|
+              # puts line if line.match(/TASK/)
+              input = line.chomp if line.match(/TASK/)
+                unless input.nil?
+                  input.each_char do |char|
+                    print "\e[34m#{char}\e[0m"
+                    $stdout.flush
+                    sleep 0.05
+                  end
+                end
+              res = line.chomp if line.match(/ok\:/)
+              puts "\e[32m#{res}\e[0m"
+            end
+        '
+        reset_vars
+    else
+        echo "Execution cancelled."
+    fi
+}
 # Export functions so they're available to subshells
 export -f extract_tasks
 export -f preview_task
@@ -158,6 +207,7 @@ export -f preview_task
 while true; do
     action=$(echo -e "Select Playbook\nSelect Host(s)\nSelect Tags\nSelect Task\nExecute Playbook\nExit" | \
         fzf --header="Ansible Menu" \
+            --layout="reverse" \
             --prompt="Action > ")
 
     case "$action" in
